pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slack-notifications"
version = "0.1.0"
description = "A simple Python library for Slack notifications at application milestones"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "slack-sdk>=3.25.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "cyclopts>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]
dev = [
    "black",
    "isort",
    "flake8",
    "mypy",
]

[project.urls]
Homepage = "https://github.com/yourusername/slack-notifications"
Repository = "https://github.com/yourusername/slack-notifications"
Issues = "https://github.com/yourusername/slack-notifications/issues"

[project.scripts]
slack-agent-cli = "slack_notifications.cli:app"
slack-mcp-server = "slack_notifications.mcp_server:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["slack_notifications*", "slack_agent*", "slack_mcp_server*"]

[tool.setuptools.package-data]
slack_notifications = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true